Crate imgchest

Structs

Body

An asynchronous request body.

Client

The client

CookieStore

An implementation for storing and retrieving [Cookie]s per the path and domain matching rules specified in RFC6265.

CookieStoreMutex

A cookie_store::CookieStore wrapped internally by a std::sync::Mutex, suitable for use in async/concurrent contexts.

CreatePostBuilder

A builder for creating a post.

FileUpdate

A file update as part of a bulk file update.

ListPostsBuilder

A builder for listing posts

ListPostsPost

The post object from a list posts call.

Post

An API post object

PostFile

An API file of a post

RawCookie

Representation of an HTTP cookie.

ScrapedPost

A Post

ScrapedPostFile

A post file

ScrapedUser

A User

UpdatePostBuilder

A builder for updating a post.

UploadPostFile

A post file that is meant for uploading.

Url

A parsed URL record.

User

The user

Enums

Error

The error

InvalidScrapedPostError

An error that may occur while parsing a post

InvalidScrapedUserError

An error that may occur while parsing a post

PostPrivacy

The post privacy

SortOrder

Sort order