Struct imgchest::Client

pub struct Client {
    pub client: Client,
    /* private fields */
}

The client

Fields
client: Client

The inner http client

Implementations

impl Client

pub fn new() -> Self

Make a new client

pub async fn get_scraped_post(&self, url: &str) -> Result<ScrapedPost, Error>

Scrape a post from a url.

Authorization

This function does NOT require the use of a token.

pub async fn load_extra_files_for_scraped_post( &self, post: &ScrapedPost, ) -> Result<Vec<ScrapedPostFile>, Error>

Load extra files for a scraped post.

Authorization

This function does NOT require the use of a token.

pub fn set_token<T>(&self, token: T)
where T: AsRef<str>,

Set the token to use for future requests.

This allows the use of functions that require authorization.

pub async fn get_post(&self, id: &str) -> Result<Post, Error>

Get a post by id.

Authorization

This function REQUIRES a token.

pub async fn create_post(&self, data: CreatePostBuilder) -> Result<Post, Error>

Create a post.

Authorization

This function REQUIRES a token.

pub async fn update_post( &self, id: &str, data: UpdatePostBuilder, ) -> Result<Post, Error>

Update a post.

Authorization

This function REQUIRES a token.

pub async fn delete_post(&self, id: &str) -> Result<(), Error>

Delete a post.

Authorization

This function REQUIRES a token.

pub async fn favorite_post(&self, id: &str) -> Result<bool, Error>

Favorite or unfavorite a post.

Returns

Returns true if the favorite was added. Returns false if the favorite was removed.

Authorization

This function REQUIRES a token.

pub async fn add_post_images<I>( &self, id: &str, images: I, ) -> Result<Post, Error>
where I: IntoIterator<Item = UploadPostFile>,

Add images to a post.

Authorization

This function REQUIRES a token.

pub async fn get_user(&self, username: &str) -> Result<User, Error>

Get a user by username.

Authorization

This function REQUIRES a token.

pub async fn get_file(&self, id: &str) -> Result<PostFile, Error>

Get a file by id.

Currently, this is implemented according to the API spec, but the API will always return no data for some reason. It is likely that this endpoint is disabled. As a result, this function is currently useless.

Authorization

This function REQUIRES a token.

pub async fn update_file( &self, id: &str, description: &str, ) -> Result<(), Error>

Update a file.

Authorization

This function REQUIRES a token.

pub async fn delete_file(&self, id: &str) -> Result<(), Error>

Delete a file.

Authorization

This function REQUIRES a token.

pub async fn update_files_bulk<I>( &self, files: I, ) -> Result<Vec<PostFile>, Error>
where I: IntoIterator<Item = FileUpdate>,

Update files in bulk.