Crate imgchest

Structs

Body

An asynchronous request body.

Client

The client

CreatePostBuilder

A builder for creating a post.

FileUpdate

A file update as part of a bulk file update.

Post

An API post object

PostFile

An API file of a post

ScrapedPost

A Post

ScrapedPostFile

A post file

UpdatePostBuilder

A builder for updating a post.

UploadPostFile

A post file that is meant for uploading.

User

The user

Enums

Error

The error

InvalidScrapedPostError

An error that may occur while parsing a post

InvalidScrapedPostFileError

Returned if a post image could not be extracted from an element

PostPrivacy

The post privacy